+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Reduce width of filter button in `flexberry-objectlistview` & `flexberry-groupedit`.
 * Now `colsconfig-dialog` has default cols width values.
 
+### Fixed
+* Fix `menuInRow` for `object-list-view`.
+* Fix inability to resize columns in `flexberry-objectlistview`, `flexberry-groupedit` and `flexberry-simpleolv` components if `tableStriped`, `rowClickable` or `customTableClass` property was changed.
+
 ## [0.8.0] - 2016-12-13
 ### Added
 * Add option `showValidationMessagesInRow` for `object-list-view`, `flexberry-objectlistview` & `flexberry-groupedit` components. Option is `false` by default in `object-list-view` & `flexberry-objectlistview`, but is `true` by default in `flexberry-groupedit`. If `true` then validation message will be shown in each cell containing invalid value.